Split apply_builders.dart.
