+++ build_runner/lib/build_runner.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'src/bootstrap/apply_builders.dart';
+import 'src/build_plan/builder_application.dart';
 import 'src/build_runner.dart' show BuildRunner;
 
 export 'src/commands/daemon/constants.dart' show assetServerPort;
