+++ build_runner/lib/src/commands/run_command.dart
@@ -11,10 +11,10 @@ import 'package:built_collection/built_collection.dart';
 import 'package:io/io.dart';
 import 'package:path/path.dart' as p;
 
-import '../bootstrap/apply_builders.dart';
 import '../bootstrap/build_process_state.dart';
 import '../build_plan/build_directory.dart';
 import '../build_plan/build_options.dart';
+import '../build_plan/builder_application.dart';
 import '../build_plan/testing_overrides.dart';
 import '../logging/build_log.dart';
 import 'build_command.dart';
