+++ build_runner/lib/src/commands/watch_command.dart
@@ -9,11 +9,11 @@ import 'package:build/experiments.dart';
 import 'package:built_collection/built_collection.dart';
 import 'package:io/io.dart';
 
-import '../bootstrap/apply_builders.dart';
 import '../bootstrap/build_process_state.dart';
 import '../build/build_result.dart';
 import '../build_plan/build_options.dart';
 import '../build_plan/build_plan.dart';
+import '../build_plan/builder_application.dart';
 import '../build_plan/testing_overrides.dart';
 import '../exceptions.dart';
 import '../logging/build_log.dart';
